pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "afem"
description = "Airframe Finite Element Modeler"
requires-python = ">=3.9,<3.13"  # Adjust based on compatibility
dependencies = [
    "pythonocc-core",
    "numpy",
    "scipy",
    "wxpython",
]
readme = {file = "README.md", content-type = "text/markdown"}
authors = [
    { name = "Trevor Laughlin", email = "info@laughlinresearch.com" },
	{ name = "Jason Corman", email = "jason.corman@gatech.edu" },
	{ name = "David Solano", email = "hsolano3@gatech.edu" }
]
license = {file = "LICENSE.txt"}
dynamic = ["version"]
keywords = ["finite element", "airframe", "CAD", "modeling"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Lesser General Public License v2 or later (LGPLv2+)",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
	"Programming Language :: Python :: 3.12",
	"Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Visualization"
]

[project.optional-dependencies]
docs = ["docs"]

[project.urls]
Repository = "https://github.com/trelau/AFEM"
Issues = "https://github.com/trelau/AFEM/issues"

[tool.setuptools]
packages = ["afem"]

[tool.setuptools.dynamic]
version = {attr = "afem.__version__"}